create direnv crate
