+++ crates/nix_health/src/check/direnv.rs
@@ -30,7 +30,7 @@ impl Checkable for Direnv {
             return checks;
         }
 
-        let direnv_install_result = direnv_crate::DirenvInstall::detect();
+        let direnv_install_result = direnv::DirenvInstall::detect();
 
         let direnv_install_check = install_check(&direnv_install_result, self.required);
         let direnv_installed = direnv_install_check.result.green();
@@ -65,10 +65,7 @@ impl Checkable for Direnv {
 }
 
 /// [Check] that direnv was installed.
-fn install_check(
-    direnv_install: &anyhow::Result<direnv_crate::DirenvInstall>,
-    required: bool,
-) -> Check {
+fn install_check(direnv_install: &anyhow::Result<direnv::DirenvInstall>, required: bool) -> Check {
     Check {
         title: "Direnv installation".to_string(),
         info: format!(
@@ -87,7 +84,7 @@ fn install_check(
 }
 
 /// [Check] that direnv version >= 2.33.0 for `direnv status --json` support
-fn version_check(direnv_install: &direnv_crate::DirenvInstall) -> Check {
+fn version_check(direnv_install: &direnv::DirenvInstall) -> Check {
     let req = VersionReq::parse(">=2.33.0").unwrap();
     let suggestion = format!("Upgrade direnv to {}", req);
     let direnv_version = &direnv_install.version;
@@ -109,7 +106,7 @@ fn version_check(direnv_install: &direnv_crate::DirenvInstall) -> Check {
 
 /// [Check] that direnv was allowed on the local flake
 fn allowed_check(
-    direnv_install: &direnv_crate::DirenvInstall,
+    direnv_install: &direnv::DirenvInstall,
     local_flake: &std::path::Path,
     required: bool,
 ) -> Check {
@@ -132,122 +129,5 @@ fn allowed_check(
             },
         },
         required,
-    }
-}
-
-/// TODO: Move this to a separate crate, called `direnv`
-/// TODO: Don't use anyhow::Result in library crates
-mod direnv_crate {
-    use semver::Version;
-    use serde::{Deserialize, Serialize};
-    use serde_repr::{Deserialize_repr, Serialize_repr};
-    use std::path::PathBuf;
-
-    /// Information about a local direnv installation
-    #[derive(Debug, Serialize, Deserialize, Clone)]
-    pub struct DirenvInstall {
-        /// Path to the direnv binary
-        pub bin_path: PathBuf,
-        /// Version of the installed direnv
-        pub version: Version,
-    }
-
-    impl DirenvInstall {
-        /// Detect user's direnv installation
-        pub fn detect() -> anyhow::Result<Self> {
-            let bin_path = which::which("direnv")?;
-            let version = get_direnv_version(&bin_path)?;
-            Ok(DirenvInstall { bin_path, version })
-        }
-
-        /// Return the `direnv status` on the given project directory
-        pub fn status(&self, project_dir: &std::path::Path) -> anyhow::Result<DirenvStatus> {
-            DirenvStatus::new(&self.bin_path, project_dir)
-        }
-    }
-
-    /// Get the version of direnv
-    fn get_direnv_version(direnv_bin: &PathBuf) -> anyhow::Result<Version> {
-        let output = std::process::Command::new(direnv_bin)
-            .args(["--version"])
-            .output()?;
-        let out = String::from_utf8_lossy(&output.stdout);
-        Ok(Version::parse(out.trim())?)
-    }
-
-    /// Information about the direnv status
-    #[derive(Debug, Serialize, Deserialize, Clone)]
-    pub struct DirenvStatus {
-        pub config: DirenvConfig,
-        pub state: DirenvState,
-    }
-
-    impl DirenvStatus {
-        /// Run `direnv status` and parse the output, for the given project directory.
-        fn new(direnv_bin: &PathBuf, dir: &std::path::Path) -> anyhow::Result<Self> {
-            let output = std::process::Command::new(direnv_bin)
-                .args(["status", "--json"])
-                .current_dir(dir)
-                .output()?;
-            let out = String::from_utf8_lossy(&output.stdout);
-            let status = DirenvStatus::from_json(&out)?;
-            Ok(status)
-        }
-
-        /// Parse the output of `direnv status --json`
-        fn from_json(json: &str) -> anyhow::Result<Self> {
-            let status: DirenvStatus = serde_json::from_str(json)?;
-            Ok(status)
-        }
-    }
-
-    #[derive(Debug, Serialize, Deserialize, Clone)]
-    pub struct DirenvConfig {
-        /// Path to the config folder of direnv
-        #[serde(rename = "ConfigDir")]
-        pub config_dir: PathBuf,
-        /// Path to the direnv binary
-        #[serde(rename = "SelfPath")]
-        pub self_path: PathBuf,
-    }
-
-    #[derive(Debug, Serialize, Deserialize, Clone)]
-    pub struct DirenvState {
-        /// Information about the .envrc found in the current directory
-        #[serde(rename = "foundRC")]
-        pub found_rc: Option<DirenvRC>,
-        /// Information about the .envrc that is currently allowed using `direnv allow`
-        #[serde(rename = "loadedRC")]
-        pub loaded_rc: Option<DirenvRC>,
-    }
-
-    impl DirenvState {
-        /// Check if the .envrc file is allowed
-        pub fn is_allowed(&self) -> bool {
-            self.found_rc
-                .as_ref()
-                .map_or(false, |rc| rc.allowed == AllowedStatus::Allowed)
-        }
-    }
-
-    /// Information about the .envrc file
-    /// TODO: Represent 0/1/2 values in a 3-value enum
-    #[derive(Debug, Serialize, Deserialize, Clone)]
-    pub struct DirenvRC {
-        pub allowed: AllowedStatus,
-        /// Path to the .envrc file
-        pub path: PathBuf,
-    }
-
-    #[derive(Serialize_repr, Deserialize_repr, PartialEq, Debug, Clone)]
-    #[repr(u8)]
-    /// Can be 0, 1 or 2
-    /// 0: Allowed
-    /// 1: NotAllowed
-    /// 2: Denied
-    pub enum AllowedStatus {
-        Allowed = 0,
-        NotAllowed = 1,
-        Denied = 2,
     }
 }
